Added Matt Stine's calculation of Pi in C.
I originally implemented this at the LA Coding Dojo in only a few minutes, but attempted to create a fancy OpenGL GUI on top of it to visualize the simulation while it ran.  Unfortunately, my LOC count exploded as I attempted to integrate Qt, get the calculation to run within the confines of Qt callbacks between drawing frames.  I never completed that part of the sim and subsequently failed to submit anything before time was up.